+++ src/builtin-services/export/MemoryMonitor.js
@@ -23,8 +23,9 @@ export class MemoryMonitor extends BaseService {
   #runner = new Threadlet(() => this.#run());
 
   /**
-   * @type {?{ heap: number, rss: number, troubleAtMsec: ?number }} Last memory
-   * snapshot (including trouble indicator), if any.
+   * @type {?{ heap: number, rss: number, at: Moment, troubleAt: ?Duration,
+   * actionAt: ?Moment }} Most recent memory snapshot (along with timing info),
+   * or `null` if a snapshot has not yet been taken.
    */
   #lastSnapshot = null;
 
@@ -71,7 +72,7 @@ export class MemoryMonitor extends BaseService {
         || (maxRssBytes  && (snapshot.rss  >= maxRssBytes))) {
       if (!snapshot.troubleAt) {
         // We just transitioned to an "over limit" situation.
-        const actionAt = now.addSec(this.config.graceSec);
+        const actionAt = now.add(this.config.gracePeriod);
         snapshot.troubleAt = now;
         snapshot.actionAt  = actionAt;
         this.logger?.overLimit({ actionAt });
@@ -93,7 +94,7 @@ export class MemoryMonitor extends BaseService {
    * Runs the service thread.
    */
   async #run() {
-    const checkMsec = this.config.checkSec * 1000;
+    const checkMsec = this.config.checkPeriod.msec;
 
     while (!this.#runner.shouldStop()) {
       const snapshot = this.#takeSnapshot();
@@ -151,11 +152,11 @@ export class MemoryMonitor extends BaseService {
    * Configuration item subclass for this (outer) class.
    */
   static #Config = class Config extends ServiceConfig {
-    /** @type {number} How often to check, in seconds. */
-    #checkSec;
+    /** @type {Duration} How often to check, in seconds. */
+    #checkPeriod;
 
-    /** @type {number} Grace period before triggering an action, in seconds. */
-    #graceSec;
+    /** @type {Duration} Grace period before triggering an action. */
+    #gracePeriod;
 
     /**
      * @type {?number} Maximum allowed size of heap usage, in bytes, or `null`
@@ -184,13 +185,13 @@ export class MemoryMonitor extends BaseService {
         maxRssBytes  = null
       } = config;
 
-      this.#checkSec = Duration.parseSec(checkPeriod ?? '5 min', { minInclusive: 1 });
-      if (this.#checkSec === null) {
+      this.#checkPeriod = Duration.parse(checkPeriod ?? '5 min', { minInclusive: 1 });
+      if (!this.#checkPeriod) {
         throw new Error(`Could not parse \`checkPeriod\`: ${checkPeriod}`);
       }
 
-      this.#graceSec = Duration.parseSec(gracePeriod ?? '0 sec', { minInclusive: 0 });
-      if (this.#graceSec === null) {
+      this.#gracePeriod = Duration.parse(gracePeriod ?? '0 sec', { minInclusive: 0 });
+      if (!this.#gracePeriod) {
         throw new Error(`Could not parse \`gracePeriod\`: ${gracePeriod}`);
       }
 
@@ -203,16 +204,14 @@ export class MemoryMonitor extends BaseService {
         : MustBe.number(maxRssBytes, { finite: true, minInclusive: 1024 * 1024 });
     }
 
-    /** @returns {number} How often to check, in seconds. */
-    get checkSec() {
-      return this.#checkSec;
+    /** @returns {Duration} How often to check. */
+    get checkPeriod() {
+      return this.#checkPeriod;
     }
 
-    /**
-     * @returns {number} Grace period before triggering an action, in seconds.
-     */
-    get graceSec() {
-      return this.#graceSec;
+    /** @returns {Duration} Grace period before triggering an action. */
+    get gracePeriod() {
+      return this.#gracePeriod;
     }
 
     /**