+++ tests/03-http2/02-not-found/run.mjs
@@ -3,6 +3,7 @@
 
 import { requestAndCheck } from '@this/integration-testing';
 
+
 await requestAndCheck(
   'Not Found',
   {